Add dashboard to package on pypi (#579)
* Move dashboard to package resource

* Update manifest with dashboard and data files

* Include word vectors as package data

* Add link to dashboard in root directory
