+++ MANIFEST.in
@@ -1,3 +1,5 @@
 include CITATION.cff
 include LICENSE
 include README.md
+recursive-include dianna/dashboard *.md *.py *.toml
+recursive-include dianna/data *.jpg *.png *.npz *.jpeg *.png *.txt *.onnx
