--- dashboard/_models_ts.py
@@ -1,5 +1,0 @@
-def predict():
-    pass
-
-
-explain_ts_dispatcher = {}
