--- dashboard/data/labels_mnist.txt
@@ -1,2 +1,0 @@
-0
-1
