--- dashboard/data/labels_text.txt
@@ -1,2 +1,0 @@
-negative
-positive
