+++ dianna/dashboard/.streamlit/config.toml
@@ -1,0 +1,10 @@
+[theme]
+base="light"
+primaryColor="7030a0"
+# backgroundColor=
+secondaryBackgroundColor="#e4f3f9"
+# textColor=
+# font=
+
+[browser]
+gatherUsageStats = false
