+++ dianna/data/labels_mnist.txt
@@ -1,0 +1,2 @@
+0
+1
