+++ dianna/data/labels_text.txt
@@ -1,0 +1,2 @@
+negative
+positive
