+++ setup.cfg
@@ -51,6 +51,8 @@ install_requires =
     # tf and tfprob are required but not declared by onnx_tf
     tensorflow >= 2.11
     tensorflow_probability
+    # https://setuptools.pypa.io/en/latest/userguide/datafiles.html#accessing-data-files-at-runtime
+    importlib_resources;python_version<'3.10'
 
 [options.data_files]
 # This section requires setuptools>=40.6.0
